refactor: extract KnowledgeSearchService from KnowledgeStoreService
Extract search operations (BM25 full-text search, hybrid retrieval with
embeddings, and triple-mode fusion) into a dedicated KnowledgeSearchService.
KnowledgeStoreService now delegates search(), findSimilarChunks(), and
searchReflections() to the new service.
